services/_shared/src/ym_shared/security.py
"""
Security utilities for YojanaMitra AI services.
"""

import bcrypt
from datetime import datetime, timedelta
from typing import Optional

import jwt
from jwt import PyJWTError


def hash_password(password: str) -> str:
    """Hash a password using bcrypt."""
    salt = bcrypt.gensalt()
    return bcrypt.hashpw(password.encode('utf-8'), salt).decode('utf-8')


def verify_password(plain_password: str, hashed_password: str) -> bool:
    """Verify a password against its hash."""
    return bcrypt.checkpw(
        plain_password.encode('utf-8'), 
        hashed_password.encode('utf-8')
    )


def create_access_token(
    data: dict, 
    secret: str, 
    algorithm: str = "HS256", 
    expires_delta: Optional[timedelta] = None
) -> str:
    """Create a JWT access token."""
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, secret, algorithm=algorithm)
    return encoded_jwt


def decode_access_token(
    token: str, 
    secret: str, 
    algorithms: list = ["HS256"]
) -> Optional[dict]:
    """Decode and validate a JWT access token."""
    try:
        payload = jwt.decode(token, secret, algorithms=algorithms)
        return payload
    except PyJWTError:
        return None
