services/_shared/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ym-shared"
version = "0.1.0"
description = "Shared utilities for YojanaMitra AI services"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "PyJWT>=2.8.0",
  "bcrypt>=4.0.0",
  "SQLAlchemy>=2.0.30",
]

[tool.setuptools.packages.find]
where = ["src"]

